requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
mangum>=0.17.0

# Data source APIs
biopython>=1.81
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0

# NLP and ML
openai>=1.3.0
langchain>=0.0.300
sentence-transformers>=2.2.0
spacy>=3.7.0

# Visualization
plotly>=5.18.0
dash>=2.14.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Database
psycopg2-binary>=2.9.0
# sqlite3 is built into Python

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0
pytz>=2023.3

